requirements.txt
# Django
Django==1.10.3
djangorestframework==3.5.3
dj-database-url==0.4.0
django-disposable-email-checker==1.2.1

# Docker
docker-compose==1.9.0rc4

# PostgresSQL
psycopg2==2.6.2

# Timezones
pytz==2016.7
django-timezone-field==2.0

# Testing
mock==2.0.0
factory-boy==2.7.0

prospector==0.12.3
bandit==1.1.0

pytest==3.0.3
pytest-cov==2.4.0
pytest-django==3.0.0

# Production
gunicorn==19.6.0